requirements.txt
streamlit
pandas
requests
scikit-learn
scipy
scikit-surprise
